requirements.txt
markitdown[all]
openai
openai-whisper
youtube_transcript_api
yt-dlp
torch
torchaudio
